Refactor configuration detail page to use new RemoveConfiguration component
